+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # ExAcme Changelog
 
+## Unreleased
+
+- Added support for handling external account binding (https://datatracker.ietf.org/doc/html/rfc8555#section-7.3.4)
+
 ## 0.3.0 (2025-03-01)
 
 - Substantial API refactoring to reduce duplication and improve maintainability.